Factor out rendering of the game to paint_game.rs
